Start querying only a subset of users
because v1’s api only allows equality tests in the where clause it can
be tricky to fetch a set of users.
use multiple queries that differ only by each ID you wish to fetch
use ID instead of _oid. Even though it is returned as _oid, the api
reads it as ID.