Update dependencies
Update dependencies and deal with the fall-out.
Fixed EJS to a version. It’s api changed so that I have to pass in the
locals as a separate parameter.
Fixed super agent usage after update to 1.x. Now the non 200 responses
return and err instead of checking res.ok
Holding moments back since the latest is incompatible with twin.