Add dynamic holiday calculation functionality
Removed hard-coded holidays--check_holiday(date) calculates if the given date is any of the simple of complex holidays specified.
Supporting logic operates under the assumption that Saturday holidays are taken on the preceding Friday, and Sunday holidays on the following Monday.

Thank you Clif for your pull request!
The last two commits together

Closes #3