Delete all code related to oauth.
Delete variables, config, classes, and tests related to the VersionOne OAuth flow.
Since the secrets file was originally from V1, and isn’t anymore, there was a lot of stuff in it that needed to move to config like .env_dev.
Only keep the actual “secret” in the secrets file. (The V1 token.)
The names of the config items used to be non-standard (i.e. lots of underscores without caps) to match the format that V1 published. Now that V1 doesn’t publish the secrets file simplify and standardize the variable names.

Closes #4