Make date query work with flagged users
The flagged users list allows you to see a much smaller list of users.
It has been broken for some time. It puked when it saw the query string that was added for carrying around a custom date range.

Make the flagged user listing tolerate and forward the custom query.
Add a link to the cadence summary directly to the flagged users list.